Renamed all tables from ASP.NET Identity. Identity classes extended.
