+++ TwitterBackup/TwitterBackup.API/Views/Account/Login.cshtml
@@ -2,7 +2,7 @@
 @using System.Linq
 @using Microsoft.AspNetCore.Http
 @using Microsoft.AspNetCore.Http.Authentication
-@using TwitterBackup.Data.Models;
+@using TwitterBackup.Data.Models.Identity;
 @model LoginViewModel
 @inject SignInManager<User> SignInManager
