+++ TwitterBackup/TwitterBackup.API/Views/Shared/_LoginPartial.cshtml
@@ -1,6 +1,6 @@
 @using Microsoft.AspNetCore.Identity
 @using TwitterBackup.API.Models
-@using TwitterBackup.Data.Models;
+@using TwitterBackup.Data.Models.Identity;
 
 @inject SignInManager<User> SignInManager
 @inject UserManager<User> UserManager
