+++ TwitterBackup/TwitterBackup.Data.Models/Identity/RoleClaim.cs
@@ -1,0 +1,8 @@
+﻿using Microsoft.AspNetCore.Identity;
+
+namespace TwitterBackup.Data.Models.Identity
+{
+    public class RoleClaim : IdentityRoleClaim<string>
+    {
+    }
+}
