+++ TwitterBackup/TwitterBackup.Data.Models/Identity/UserClaim.cs
@@ -1,0 +1,8 @@
+﻿using Microsoft.AspNetCore.Identity;
+
+namespace TwitterBackup.Data.Models.Identity
+{
+    public class UserClaim : IdentityUserClaim<string>
+    {
+    }
+}
