+++ TwitterBackup/TwitterBackup.Data.Models/Identity/UserRole.cs
@@ -1,0 +1,8 @@
+﻿using Microsoft.AspNetCore.Identity;
+
+namespace TwitterBackup.Data.Models.Identity
+{
+    public class UserRole : IdentityUserRole<string>
+    {
+    }
+}
